helper: Fixing bug with ndn::GlobalRoutingHelper
With certain metric parameters, the value was incorrectly wrapped
around zero, leading to incorrect route calculation (BoostGrapLibrary
was detecting negative edge weights).

+++ model/ndn-face.h
@@ -273,7 +273,7 @@ private:
   DataHandler m_upstreamDataHandler;
   bool m_ifup;
   uint32_t m_id; ///< \brief id of the interface in NDN stack (per-node uniqueness)
-  uint32_t m_metric; ///< \brief metric of the face
+  uint16_t m_metric; ///< \brief metric of the face
   uint32_t m_flags; ///< @brief faces flags (e.g., APPLICATION)
 };
